// indexer/src/types/enums.ts
// SPDX-License-Identifier: Apache-2.0

// Auto-generated , DO NOT EDIT

export enum TransactionType {

    SUBSTRATE = "SUBSTRATE",

    EVM = "EVM",

    EVM_WRAPPED = "EVM_WRAPPED",

}


export enum TransactionStatus {

    SUCCESS = "SUCCESS",

    FAILED = "FAILED",

    PENDING = "PENDING",

}


export enum TokenStandard {

    ERC20 = "ERC20",

    ERC721 = "ERC721",

    ERC1155 = "ERC1155",

    PSP22 = "PSP22",

    PSP34 = "PSP34",

    PSP37 = "PSP37",

    NATIVE = "NATIVE",

}


export enum ContractType {

    EVM = "EVM",

    INK = "INK",

}


export enum StakingStatus {

    ACTIVE = "ACTIVE",

    WAITING = "WAITING",

    NOMINATING = "NOMINATING",

    IDLE = "IDLE",

    CHILLED = "CHILLED",

}


export enum ProposalStatus {

    ACTIVE = "ACTIVE",

    PASSED = "PASSED",

    REJECTED = "REJECTED",

    CANCELLED = "CANCELLED",

    EXECUTED = "EXECUTED",

    TIMEOUT = "TIMEOUT",

}


export enum VoteType {

    AYE = "AYE",

    NAY = "NAY",

    ABSTAIN = "ABSTAIN",

}